Move theme classes to their own file
